requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.11
PyYAML>=6.0
yfinance>=0.2.40
torch>=2.2
pyarrow>=14.0
